backend/pyproject.toml
[project]
name = "annotation-backend"
version = "0.1.0"
description = "Backend for the annotation project"
requires-python = ">=3.8"
dependencies = [
    "Django==5.0.2",
    "django-filter==24.1",
    "django-polymorphic==3.1.0",
    "djangorestframework==3.14.0",
    "django-cors-headers==4.3.1",
    "python-dotenv==1.0.1",
    "uwsgi==2.0.24",
]
